requirements.txt
python-dotenv
openai
anthropic